фикс карты

+++ server/src/main/groovy/com/riversoft/game/snake/model/RoundInfo.groovy
@@ -8,5 +8,5 @@ class RoundInfo {
 }
 class RoundPlayerInfo{
     String name;
-    int kpd;
+    float kpd;
 }
